+++ core/minute-ia/atomic.h
@@ -11,7 +11,7 @@
 #include "common.h"
 #include "util.h"
 
-typedef int atomic_t;
+typedef long atomic_t;
 typedef atomic_t atomic_val_t;
 
 static inline int bool_compare_and_swap_u32(uint32_t *var, uint32_t old_value,
